+++ src/Exception/JsonException.php
@@ -3,7 +3,7 @@
  * PHP simple library for managing Json files.
  *
  * @author    Josantonius <hello@josantonius.com>
- * @copyright 2016 - 2017 (c) Josantonius - PHP-Json
+ * @copyright 2016 - 2018 (c) Josantonius - PHP-Json
  * @license   https://opensource.org/licenses/MIT - The MIT License (MIT)
  * @link      https://github.com/Josantonius/PHP-Json
  * @since     1.0.0
@@ -15,16 +15,12 @@ namespace Josantonius\Json\Exception;
  *
  * You can use an exception and error handler with this library.
  *
- * @since 1.0.0
- *
  * @link https://github.com/Josantonius/PHP-ErrorHandler
  */
 class JsonException extends \Exception
 {
     /**
      * Exception handler.
-     *
-     * @since 1.0.0
      *
      * @param string $msg    → message error (Optional)
      * @param int    $status → HTTP response status code (Optional)