+++ src/Json.php
@@ -3,7 +3,7 @@
  * PHP simple library for managing Json files.
  *
  * @author    Josantonius <hello@josantonius.com>
- * @copyright 2016 - 2017 (c) Josantonius - PHP-Json
+ * @copyright 2016 - 2018 (c) Josantonius - PHP-Json
  * @license   https://opensource.org/licenses/MIT - The MIT License (MIT)
  * @link      https://github.com/Josantonius/PHP-Json
  * @since     1.0.0
@@ -11,42 +11,35 @@
 namespace Josantonius\Json;
 
 use Josantonius\Json\Exception\JsonException;
-use Josantonius\Json\Exception\JsonLastErrorException;
 
 /**
  * Json handler.
- *
- * @since 1.0.0
  */
 class Json
 {
     /**
      * Creating JSON file from array.
      *
-     * @since 1.0.0
-     *
      * @param array  $array → array to be converted to JSON
      * @param string $file  → path to the file
      *
-     * @return bool → true if the file is created without errors
+     * @return boolean → true if the file is created without errors
      */
     public static function arrayToFile($array, $file)
     {
         self::createDirectory($file);
 
-        $json = json_encode($array, JSON_PRETTY_PRINT);
+        $lastError = JsonLastError::check();
 
-        $json = self::jsonLastError() ? json_encode($json, 128) : $json;
+        $json = json_encode($lastError ? $lastError : $array, JSON_PRETTY_PRINT);
 
         self::saveFile($file, $json);
 
-        return ! isset($json['error-code']);
+        return is_null($lastError);
     }
 
     /**
      * Save to array the JSON file content.
-     *
-     * @since 1.0.0
      *
      * @param string $file → path or external url to JSON file
      *
@@ -58,11 +51,11 @@ class Json
             self::arrayToFile([], $file);
         }
 
-        $jsonString = @file_get_contents($file);
-        $array = json_decode($jsonString, true);
-        $error = self::jsonLastError();
+        $json = @file_get_contents($file);
+        $array = json_decode($json, true);
+        $lastError = JsonLastError::check();
 
-        return $array === null || isset($error['error-code']) ? false : $array;
+        return $array === null || !is_null($lastError) ? false : $array;
     }
 
     /**
@@ -103,84 +96,5 @@ class Json
             $message = 'Could not create file in';
             throw new JsonException($message . ' ' . $file);
         }
-    }
-
-    /**
-     * The JSON last error collections.
-     *
-     * @since 1.1.3
-     *
-     * @return array
-     */
-    private static function jsonLastErrorCollections()
-    {
-        $collections = [
-            JSON_ERROR_NONE => null,
-            JSON_ERROR_DEPTH => [
-                'message' => 'Maximum stack depth exceeded',
-                'error-code' => 1,
-            ],
-            JSON_ERROR_STATE_MISMATCH => [
-                'message' => 'Underflow or the modes mismatch',
-                'error-code' => 2,
-            ],
-            JSON_ERROR_CTRL_CHAR => [
-                'message' => 'Unexpected control char found',
-                'error-code' => 3,
-            ],
-            JSON_ERROR_SYNTAX => [
-                'message' => 'Syntax error, malformed JSON',
-                'error-code' => 4,
-            ],
-            JSON_ERROR_UTF8 => [
-                'message' => 'Malformed UTF-8 characters',
-                'error-code' => 5,
-            ],
-            JSON_ERROR_RECURSION => [
-                'message' => 'Recursion error in value to be encoded',
-                'error-code' => 6,
-            ],
-            JSON_ERROR_INF_OR_NAN => [
-                'message' => 'Error NAN/INF in value to be encoded',
-                'error-code' => 7,
-            ],
-            JSON_ERROR_UNSUPPORTED_TYPE => [
-                'message' => 'Type value given cannot be encoded',
-                'error-code' => 8,
-            ],
-            'default' => [
-                'message' => 'Unknown error',
-                'error-code' => 999,
-            ],
-        ];
-
-        if (version_compare(PHP_VERSION, '7.0.0', '>='))
-        {
-            $collections[JSON_ERROR_INVALID_PROPERTY_NAME] = [
-                'message' => 'Name value given cannot be encoded',
-                'error-code' => 9,
-            ];
-            $collections[JSON_ERROR_UTF16] = [
-                'message' => 'Malformed UTF-16 characters',
-                'error-code' => 10,
-            ];
-        }
-
-        return $collections;
-    }
-
-    /**
-     * Check for errors.
-     *
-     * @since 1.1.3
-     *
-     * @return array|null
-     */
-    private static function jsonLastError()
-    {
-        $collections = self::jsonLastErrorCollections();
-        $jsonLastError = json_last_error();
-
-        return isset($jsonLastError) ? $collections[$jsonLastError] : $collections['default'];
     }
 }